+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Summary of release changes for Version 1 - CentOS-6
 
 - Fixes bootstrap errors regarding readonly `PASSWORD_LENGTH`.
 - Updates validation for `SSH_USER_ID` to allow values in the system ID range.
+- Adds `SSH_USER_PRIVATE_KEY` to allow configuration of an RSA private key for `SSH_USER`.
 
 ### 1.9.1 - 2018-11-10